Refactor : 디렉토리 구조 변경
더 직관적인 디렉토리 구조를 위해 Cam/Menu 디렉토리를 생성하여 ButtonBar, ChattingTab, UserListTab 관련 모듈을 이동하였습니다.
